LayoutTests/http/wpt/web-animations/timing-model/time-transformations/transformed-progress-expected.txt


FAIL Transformed progress for step-start function assert_approx_equals: The progress should be approximately 1 at 0ms expected a number but got a "object"
FAIL Transformed progress for steps(1, start) function assert_approx_equals: The progress should be approximately 1 at 0ms expected a number but got a "object"
FAIL Transformed progress for steps(2, start) function assert_approx_equals: The progress should be approximately 0.5 at 0ms expected a number but got a "object"
FAIL Transformed progress for step-end function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for steps(1) function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for steps(1, end) function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for steps(2, end) function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for frames function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for linear function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for ease function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for ease-in function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for ease-in-out function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for ease-out function assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for easing function which produces values greater than 1 assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Transformed progress for easing function which produces values less than 1 assert_approx_equals: The progress should be approximately 0 at 0ms expected a number but got a "object"
FAIL Test bounds point of step-start easing assert_equals: Progress at 0ms expected (number) 0 but got (object) null
FAIL Test bounds point of step-start easing with reverse direction assert_equals: Progress at 0ms expected (number) 1 but got (object) null
FAIL Test bounds point of step-start easing with iterationStart not at a transition point assert_equals: Progress at 0ms expected (number) 0.5 but got (object) null
FAIL Test bounds point of step-start easing with iterationStart and delay assert_equals: Progress at 0ms expected (number) 0.5 but got (object) null
FAIL Test bounds point of step-start easing with iterationStart and reverse direction assert_equals: Progress at 0ms expected (number) 1 but got (object) null
FAIL Test bounds point of step(4, start) easing with iterationStart 0.75 and delay assert_equals: Progress at 0ms expected (number) 0.75 but got (object) null
FAIL Test bounds point of step-start easing with alternate direction assert_equals: Progress at 0ms expected (number) 1 but got (object) null
FAIL Test bounds point of step-start easing with alternate-reverse direction assert_equals: Progress at 0ms expected (number) 1 but got (object) null
FAIL Test bounds point of step-end easing assert_equals: Progress at 0ms expected (number) 0 but got (object) null
FAIL Test bounds point of step-end easing with iterationStart and delay assert_equals: Progress at 0ms expected (number) 0 but got (object) null
FAIL Test bounds point of step-end easing with iterationStart not at a transition point assert_equals: Progress at 0ms expected (number) 0.5 but got (object) null
FAIL Test bounds point of frames easing assert_equals: Progress at 0ms expected (number) 0 but got (object) null
FAIL Test bounds point of frames easing with iterationStart and delay assert_equals: Progress at 0ms expected (number) 1 but got (object) null
